refactor(socket.io): deprecate categories.categorySearch in favour of api.search.categories
THIS IS WIP
