+++ src/api/index.js
@@ -9,6 +9,7 @@ module.exports = {
 	posts: require('./posts'),
 	chats: require('./chats'),
 	categories: require('./categories'),
+	search: require('./search'),
 	flags: require('./flags'),
 	files: require('./files'),
 	utils: require('./utils'),
