+++ src/routes/write/index.js
@@ -41,6 +41,7 @@ Write.reload = async (params) => {
 	router.use('/api/v3/posts', require('./posts')());
 	router.use('/api/v3/chats', require('./chats')());
 	router.use('/api/v3/flags', require('./flags')());
+	router.use('/api/v3/search', require('./search')());
 	router.use('/api/v3/admin', require('./admin')());
 	router.use('/api/v3/files', require('./files')());
 	router.use('/api/v3/utilities', require('./utilities')());
